moved android and wind32 window implementations into platform folder
