+++ include/vsg/platform/win32/Win32_Window.h
